Move the vocabulary-only files from `src/index` into `src/index/vocabulary`
`Vocabulary.{h,cpp}`, `EncodedIriManager.{h,cpp}`, and `PrefixHeuristic.{h,cpp}`
used to live directly in `src/index` and were compiled into the `index` library,
although all three of them are purely vocabulary concerns and none of them uses a
single symbol of the `index` library. `Vocabulary.cpp` resolves completely inside
the `vocabulary` library together with its existing `qlever_util rdfTypes` link
line, and `PrefixHeuristic` is only used by `CompressedVocabulary.h` and
`CompressionWrappers.h`, which both already live in `src/index/vocabulary`.

This commit moves the six files into `src/index/vocabulary`, adds the three
`.cpp` files to the `vocabulary` target, and removes them from the `index`
target. The include guards of the moved headers are adapted to their new paths;
note that the guard of `EncodedIriManager.h` was `QLEVER_SRC_INDEX_ENCODEDVALUES_H`
before, which did not even match its old filename.

The same argument applies to `StringSortComparator.h` and `LocaleManager.h`,
which are moved as well. Both are header-only and are already included from
`src/index/vocabulary`, so there is no `.cpp` file for either of them and hence
no CMake change at all; only their include guards and the references to them (in
`#include` directives as well as in comments) are updated. The guard of
`StringSortComparator.h` was `QLEVER_STRINGSORTCOMPARATOR_H` before, which did
not follow the naming scheme of the other guards at all.

In total, 38 include sites in `src` and `test` are updated to the new paths.

Apart from being the right place for these files, the move also removes an
existing layering violation: `src/index/vocabulary/SplitVocabularyImpl.h`
included `index/Vocabulary.h`, i.e. the `vocabulary` library reached back into
`src/index`.

Finally, four tests no longer need the heavy `index` or `parser` libraries and
now link `vocabulary` instead: `EncodedIriManagerTest`, `VocabularyOnDiskTest`,
`SplitVocabularyTest`, and `VocabularyTest`.
