+++ src/VocabularyMergerMain.cpp
@@ -6,8 +6,8 @@
 // Can be used e.g. for benchmarking this step to develop faster IndexBuilders.
 
 #include "global/FileSuffixConstants.h"
-#include "index/Vocabulary.h"
 #include "index/VocabularyMerger.h"
+#include "index/vocabulary/Vocabulary.h"
 
 // ____________________________________________________________________________________________________
 int main(int argc, char** argv) {
