+++ src/engine/ExportQueryExecutionTrees.cpp
@@ -24,9 +24,9 @@
 #include "backports/algorithm.h"
 #include "engine/ConstructTripleGenerator.h"
 #include "global/RuntimeParameters.h"
-#include "index/EncodedIriManager.h"
 #include "index/ExportIds.h"
 #include "index/IndexImpl.h"
+#include "index/vocabulary/EncodedIriManager.h"
 #include "rdfTypes/RdfEscaping.h"
 #include "util/ConstexprUtils.h"
 #include "util/ValueIdentity.h"
