+++ src/index/Index.h
@@ -15,10 +15,10 @@
 #include "index/GraphNameManager.h"
 #include "index/InputFileSpecification.h"
 #include "index/Permutation.h"
-#include "index/StringSortComparator.h"
 #include "index/TextScanMode.h"
 #include "index/TextScoringEnum.h"
-#include "index/Vocabulary.h"
+#include "index/vocabulary/StringSortComparator.h"
+#include "index/vocabulary/Vocabulary.h"
 #include "parser/TripleComponent.h"
 #include "util/CancellationHandle.h"
 #include "util/json.h"
