+++ src/index/IndexImpl.h
@@ -25,7 +25,6 @@
 #include "index/ConstantsIndexBuilding.h"
 #include "index/DeltaTriples.h"
 #include "index/DocsDB.h"
-#include "index/EncodedIriManager.h"
 #include "index/ExternalSortFunctors.h"
 #include "index/GraphNameManager.h"
 #include "index/Index.h"
@@ -35,8 +34,9 @@
 #include "index/Permutation.h"
 #include "index/TextMetaData.h"
 #include "index/TextScoring.h"
-#include "index/Vocabulary.h"
 #include "index/VocabularyMerger.h"
+#include "index/vocabulary/EncodedIriManager.h"
+#include "index/vocabulary/Vocabulary.h"
 #include "parser/RdfParser.h"
 #include "parser/TripleComponent.h"
 #include "util/File.h"
