+++ src/index/vocabulary/LocaleManager.h
@@ -8,8 +8,8 @@
 // You may not use this file except in compliance with the Apache 2.0 License,
 // which can be found in the `LICENSE` file at the root of the QLever project.
 
-#ifndef QLEVER_SRC_INDEX_LOCALEMANAGER_H
-#define QLEVER_SRC_INDEX_LOCALEMANAGER_H
+#ifndef QLEVER_SRC_INDEX_VOCABULARY_LOCALEMANAGER_H
+#define QLEVER_SRC_INDEX_VOCABULARY_LOCALEMANAGER_H
 
 #ifndef QLEVER_NO_UNICODE
 #include <unicode/coll.h>
@@ -433,4 +433,4 @@ using LocaleManager = LocaleManagerNoICU;
 using LocaleManager = LocaleManagerICU;
 #endif
 
-#endif  // QLEVER_SRC_INDEX_LOCALEMANAGER_H
+#endif  // QLEVER_SRC_INDEX_VOCABULARY_LOCALEMANAGER_H
