+++ src/index/vocabulary/PrefixHeuristic.h
@@ -2,8 +2,8 @@
 // Chair of Algorithms and Data Structures.
 // Author: Johannes Kalmbach<joka921> (johannes.kalmbach@gmail.com)
 
-#ifndef QLEVER_SRC_INDEX_PREFIXHEURISTIC_H
-#define QLEVER_SRC_INDEX_PREFIXHEURISTIC_H
+#ifndef QLEVER_SRC_INDEX_VOCABULARY_PREFIXHEURISTIC_H
+#define QLEVER_SRC_INDEX_VOCABULARY_PREFIXHEURISTIC_H
 
 #include <memory>
 #include <string>
@@ -130,4 +130,4 @@ class Tree {
 
 }  // namespace ad_utility
 
-#endif  // QLEVER_SRC_INDEX_PREFIXHEURISTIC_H
+#endif  // QLEVER_SRC_INDEX_VOCABULARY_PREFIXHEURISTIC_H
