+++ src/index/vocabulary/SplitVocabularyImpl.h
@@ -6,9 +6,9 @@
 #define QLEVER_SRC_INDEX_VOCABULARY_SPLITVOCABULARYIMPL_H
 
 #include "backports/type_traits.h"
-#include "index/Vocabulary.h"
 #include "index/vocabulary/GeoVocabulary.h"
 #include "index/vocabulary/SplitVocabulary.h"
+#include "index/vocabulary/Vocabulary.h"
 #include "util/Log.h"
 #include "util/TypeTraits.h"
