+++ src/parser/RdfParser.cpp
@@ -21,8 +21,8 @@
 #include "backports/StartsWithAndEndsWith.h"
 #include "engine/CallFixedSize.h"
 #include "global/Constants.h"
-#include "index/EncodedIriManager.h"
 #include "index/InputFileSpecification.h"
+#include "index/vocabulary/EncodedIriManager.h"
 #include "parser/NormalizedString.h"
 #include "parser/Tokenizer.h"
 #include "parser/TokenizerCtre.h"
