+++ src/parser/RdfParser.h
@@ -23,8 +23,8 @@
 #include "global/Constants.h"
 #include "global/SpecialIds.h"
 #include "index/ConstantsIndexBuilding.h"
-#include "index/EncodedIriManager.h"
 #include "index/InputFileSpecification.h"
+#include "index/vocabulary/EncodedIriManager.h"
 #include "parser/AsyncBlockSource.h"
 #include "parser/AsyncFileBlockDriver.h"
 #include "parser/TripleComponent.h"
