+++ src/parser/TripleComponent.h
@@ -16,8 +16,8 @@
 #include "global/Constants.h"
 #include "global/Id.h"
 #include "global/SpecialIds.h"
-#include "index/EncodedIriManager.h"
 #include "index/LocalVocab.h"
+#include "index/vocabulary/EncodedIriManager.h"
 #include "parser/LiteralOrIri.h"
 #include "rdfTypes/RdfEscaping.h"
 #include "rdfTypes/Variable.h"
