+++ src/parser/data/GraphTerm.h
@@ -11,7 +11,7 @@
 #include <variant>
 
 #include "backports/StartsWithAndEndsWith.h"
-#include "index/EncodedIriManager.h"
+#include "index/vocabulary/EncodedIriManager.h"
 #include "parser/RdfParser.h"
 #include "parser/TokenizerCtre.h"
 #include "parser/data/BlankNode.h"
