+++ test/TripleComponentTest.cpp
@@ -8,7 +8,7 @@
 #include "./util/IdTestHelpers.h"
 #include "./util/TripleComponentTestHelpers.h"
 #include "global/ValueId.h"
-#include "index/EncodedIriManager.h"
+#include "index/vocabulary/EncodedIriManager.h"
 #include "parser/TripleComponent.h"
 #include "rdfTypes/GeoPoint.h"
 #include "rdfTypes/Literal.h"
