+++ test/ValueIdTest.cpp
@@ -13,8 +13,8 @@
 #include "./util/IndexTestHelpers.h"
 #include "backports/algorithm.h"
 #include "global/ValueId.h"
-#include "index/EncodedIriManager.h"
 #include "index/LocalVocabEntry.h"
+#include "index/vocabulary/EncodedIriManager.h"
 #include "util/HashSet.h"
 #include "util/Random.h"
 #include "util/Serializer/ByteBufferSerializer.h"
