+++ test/VocabularyGeneratorTest.cpp
@@ -20,10 +20,10 @@
 #include "global/Constants.h"
 #include "index/ConstantsIndexBuilding.h"
 #include "index/Index.h"
-#include "index/Vocabulary.h"
 #include "index/VocabularyMerger.h"
 #include "index/vocabulary/CompressedVocabulary.h"
 #include "index/vocabulary/SplitVocabulary.h"
+#include "index/vocabulary/Vocabulary.h"
 #include "index/vocabulary/VocabularyInternalExternal.h"
 #include "util/Algorithm.h"
 #include "util/File.h"
