+++ test/index/CMakeLists.txt
@@ -2,7 +2,7 @@ add_subdirectory(vocabulary)
 addLinkAndDiscoverTest(PatternCreatorTest index)
 addLinkAndDiscoverTest(ScanSpecificationTest index)
 addLinkAndDiscoverTestNoLibs(KeyOrderTest)
-addLinkAndDiscoverTestNoLibs(EncodedIriManagerTest parser)
+addLinkAndDiscoverTestNoLibs(EncodedIriManagerTest vocabulary)
 addLinkAndDiscoverTest(GraphNameManagerTest index)
 addLinkAndDiscoverTest(IndexRebuilderTest index server)
 addLinkAndDiscoverTest(InputFileSpecificationTest parser index)
