+++ test/index/vocabulary/GeoVocabularyTest.cpp
@@ -7,9 +7,9 @@
 #include "../../GeometryInfoTestHelpers.h"
 #include "VocabularyTestHelpers.h"
 #include "gmock/gmock.h"
-#include "index/Vocabulary.h"
 #include "index/vocabulary/CompressedVocabulary.h"
 #include "index/vocabulary/GeoVocabulary.h"
+#include "index/vocabulary/Vocabulary.h"
 #include "index/vocabulary/VocabularyInMemory.h"
 #include "index/vocabulary/VocabularyInternalExternal.h"
 #include "util/File.h"
